Move User model class to App\Models namespace
